\name{letters_out}
\docType{data}
\alias{letters_out}

\title{Letters training sample- output}

\description{This data set gives the expected response for the \code{letters_train} dataset. Each letter has more than one representation. }

\usage{data(letters)}
\format{A matrix containing 40 response. Each row represents one expected response.}
\source{Collecting from about 10 person and from some frequently used font type}


\keyword{datasets}