\name{letters_recall}
\docType{data}
\alias{letters_recall}

\title{Letters recall sample- input}

\description{This data set gives the 8x8 bitmatrix representation of alphabetic letters from A to E. Each letter has more than one representation. Can be useful as a recalling set for mlp network. }

\usage{data(letters_recall)}
\format{A matrix containing 50 representation of letters. Each row represents one letter. }
\source{Collecting from about 10 person and from some frequently used font type}

\keyword{datasets}