\name{letters_train}
\docType{data}
\alias{letters_train}

\title{Letters training sample- input}

\description{This data set gives the 8x8 bitmatrix representation of alphabetic letters from A to E. Each letter has more than one representation. Can be useful as an input training set for mlp network. }

\usage{data(letters)}
\format{A matrix containing 40 representation of letters. Each row represents one letter. }
\source{Collecting from about 10 person and from some frequently used font type}

\keyword{datasets}