% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{link}
\alias{link}
\title{link function (to output layer)}
\usage{
link(f, df = NULL, ...)
}
\arguments{
\item{f}{link function to use (right now only use alt_softmax)}

\item{df}{derivative of link function}
}
\description{
link function transforms the last hidden layer into
required output format without adding any additional parameters.
e.g. softmax (the only link function used at the moment)
}
