% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{loss}
\alias{loss}
\title{loss function}
\usage{
loss(f, df = NULL, ...)
}
\arguments{
\item{f}{loss function. should be a function of \code{y} (target value) and \code{ypred} (predicted value)}

\item{df}{derivative of loss function. Ideally, \code{f} is differentiable by \code{\link[Deriv:Deriv]{Deriv::Deriv()}} so you do not need to manually specify the derivative.}
}
\description{
loss function
}
