% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_estimate.R
\name{nn_estimate}
\alias{nn_estimate}
\title{train a neural network}
\usage{
nn_estimate(nnobject, learnrate, gradmethod, nbatch, nepoch, ...)
}
\arguments{
\item{nnobject}{a list of class "nn" containing the inputs gathered from each layer.}

\item{learnrate}{learning rate for gradient descent}

\item{gradmethod}{gradient descent method (currently not used)}

\item{nbatch}{randomly divides training examples into \code{nbatch} batches while training}

\item{nepoch}{number of passes through all training examples}
}
\description{
This function is the workhorse of the package, doing the actual training of the model.
You should not ever have to use \code{nn_estimate} directly.
}
