% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{penalty}
\alias{penalty}
\title{penalty function}
\usage{
penalty(f, df = NULL, lambda = 0, ...)
}
\arguments{
\item{f}{loss function. should be a vectorized function of a single argument \code{x}}

\item{df}{derivative of loss function. Ideally, \code{f} is differentiable by \code{\link[Deriv:Deriv]{Deriv::Deriv()}} so you do not need to manually specify the derivative.}

\item{lambda}{(non-negative) coefficient on the penalty term (bigger means more penalized)}
}
\description{
adds regularization term to the loss function
}
