% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{+.nn}
\alias{+.nn}
\title{specify neural network layers}
\usage{
\method{+}{nn}(o1, o2)
}
\description{
This function overloads \code{+} so that neural network layers can be specified
analogous to how \code{ggplot} syntax works.
}
