% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_functions.R
\name{train}
\alias{train}
\title{train neural network}
\usage{
train(learnrate, gradmethod = "gradient descent", nbatch = 1, nepoch, ...)
}
\arguments{
\item{learnrate}{learning rate for gradient descent}

\item{gradmethod}{gradient descent method (currently not used)}

\item{nbatch}{randomly divides training examples into \code{nbatch} batches while training}

\item{nepoch}{number of passes through all training examples}
}
\description{
This function specifies training parameters.
}
